requirement.txt
absl-py==2.3.0
altair==5.5.0
astunparse==1.6.3
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
device==0.1
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.58.4
gast==0.4.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.40.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.73.1
h5py==3.14.0
idna==3.10
importlib-metadata==6.11.0
importlib_resources==6.5.2
jax==0.4.30
jaxlib==0.4.30
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
keras==2.12.0
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
narwhals==1.44.0
networkx==3.2.1
numpy==1.23.5
oauthlib==3.3.1
opt_einsum==3.4.0
packaging==23.2
pandas==2.3.0
Pillow==9.5.0
protobuf==4.25.8
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.36.2
requests==2.0.0
requests-oauthlib==2.0.0
rich==13.9.4
rpds-py==0.25.1
rsa==4.9.1
scipy==1.13.1
six==1.17.0
smmap==5.0.2
streamlit==1.27.0
sympy==1.14.0
tenacity==8.5.0
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
toml==0.10.2
torch==2.0.1
torchaudio==2.0.2
torchsummary==1.5.1
torchvision==0.15.2
tornado==6.5.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
validators==0.35.0
watchdog==6.0.0
Werkzeug==3.1.3
wrapt==1.14.1
zipp==3.23.0
